Reworked systems to be simpler, created more razor components, and added upgrades and multiplier upgrades.

+++ IdleGame.Domain/IResource.cs
@@ -5,6 +5,6 @@
         string Name { get; }
         long Quantity { get; }
 
-        void Add(int quantity);
+        void Add(long quantity);
     }
 }
